Split errors
